# Remove `OPTIMIZER_FEATURES_ENABLE('11.2.0.2')` hint