This pull request removes `OPTIMIZER_FEATURES_ENABLE('11.2.0.2')` hint

This hint caused slowness #2179, and I decided to remove this hint.

There are no significant elapsed time changes with/without this hint in our CI:

* With `OPTIMIZER_FEATURES_ENABLE('11.2.0.2')` hint

- test that runs Oracle Database 23c XE
https://github.com/rsim/oracle-enhanced/actions/runs/4963761266/jobs/8883291198?pr=2337

> Finished in 5 minutes 43 seconds (files took 0.70408 seconds to load)

- test_11g that runs Oracle Database 11g XE
https://github.com/rsim/oracle-enhanced/actions/runs/4963761261/jobs/8883291209?pr=2337

> Finished in 17 minutes 39 seconds (files took 0.6625 seconds to load)

* Without `OPTIMIZER_FEATURES_ENABLE('11.2.0.2')` hint

- test that runs Oracle Database 23c XE
https://github.com/rsim/oracle-enhanced/actions/runs/4963590072/jobs/8882971220?pr=2337

> Finished in 5 minutes 45 seconds (files took 0.67993 seconds to load)

- test_11g that runs Oracle Database 11g XE
https://github.com/rsim/oracle-enhanced/actions/runs/4963590070/jobs/8882971277?pr=2337

> Finished in 17 minutes 43 seconds (files took 0.7018 seconds to load)

Fix #2179