@yahonda 2023-05-12
- test / build (3.2)
https://github.com/rsim/oracle-enhanced/actions/runs/4963590072/jobs/8882971220?pr=2337

> Finished in 5 minutes 45 seconds (files took 0.67993 seconds to load)

- test_11g / build (3.2)
https://github.com/rsim/oracle-enhanced/actions/runs/4963590070/jobs/8882971277?pr=2337

> Finished in 17 minutes 43 seconds (files took 0.7018 seconds to load)