@yahonda 2023-05-13
Related to 
https://github.com/rsim/oracle-enhanced/issues/2179